+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [Unreleased][unreleased]
 
-## Changed
-
+### Changed
+- new webxdc api signature for `Context.getWebxdcStatusUpdates`
 - Upgrade to deltachat-core version `1.77.0`
 
 ## [1.76.0] - 2022-03-04