+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+## Removed
+- `createTempUser()`
+
 ### Changed
 - new webxdc api signature for `Context.getWebxdcStatusUpdates`
 - Upgrade to deltachat-core version `1.77.0`